pyproject.toml
[project]
name = "jcli"
version = "0.1.0"
description = "CLI tool for Juniper Junos device management"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "junos-eznc>=2.7.4",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "jxmlease>=1.0.3",
]

[project.scripts]
jcli = "jcli.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jcli"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "ruff>=0.14.0",
]
